successfully refactored imports and solvers so everything is cleaner now.
