+++ pulp/__init__.py
@@ -33,5 +33,9 @@ from .constants import VERSION
 
 from .pulp import *
 from .amply import *
+from .apis import *
+from .utilities import *
+from .constants import *
+
 __doc__ = pulp.__doc__
 __version__ = VERSION
