+++ tests/test_pulp.py
@@ -1,7 +1,10 @@
 """
 Tests for pulp
 """
-from pulp import *
+from pulp.constants import PulpError
+from pulp.apis import *
+from pulp import LpVariable, LpProblem, lpSum, LpConstraintVar, LpFractionConstraint
+from pulp import constants as const
 import unittest
 
 
@@ -56,7 +59,7 @@ class PuLPTest(unittest.TestCase):
 
     def test_pulp_009(self):
         # infeasible
-        prob = LpProblem("test09", LpMinimize)
+        prob = LpProblem("test09", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -69,18 +72,18 @@ class PuLPTest(unittest.TestCase):
         print("\t Testing inconsistant lp solution")
         # this was a problem with use_mps=false
         if self.solver.__class__ in [PULP_CBC_CMD, COIN_CMD]:
-            pulpTestCheck(prob, self.solver, [LpStatusInfeasible], {x: 4, y: -1, z: 6, w: 0},
+            pulpTestCheck(prob, self.solver, [const.LpStatusInfeasible], {x: 4, y: -1, z: 6, w: 0},
                           use_mps=False)
         elif self.solver.__class__ in [PULP_CHOCO_CMD, CHOCO_CMD, MIPCL_CMD]:
             # this error is not detected with mps and choco, MIPCL_CMD can only use mps files
             pass
         else:
-            pulpTestCheck(prob, self.solver, [LpStatusInfeasible, LpStatusNotSolved,
-                                              LpStatusUndefined])
+            pulpTestCheck(prob, self.solver, [const.LpStatusInfeasible, const.LpStatusNotSolved,
+                                              const.LpStatusUndefined])
 
     def test_pulp_010(self):
         # Continuous
-        prob = LpProblem("test010", LpMinimize)
+        prob = LpProblem("test010", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -91,11 +94,11 @@ class PuLPTest(unittest.TestCase):
         prob += -y + z == 7, "c3"
         prob += w >= 0, "c4"
         print("\t Testing continuous LP solution")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
 
     def test_pulp_011(self):
         # Continuous Maximisation
-        prob = LpProblem("test011", LpMaximize)
+        prob = LpProblem("test011", const.LpMaximize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -106,11 +109,11 @@ class PuLPTest(unittest.TestCase):
         prob += -y + z == 7, "c3"
         prob += w >= 0, "c4"
         print("\t Testing maximize continuous LP solution")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: 1, z: 8, w: 0})
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: 1, z: 8, w: 0})
 
     def test_pulp_012(self):
         # Unbounded
-        prob = LpProblem("test012", LpMaximize)
+        prob = LpProblem("test012", const.LpMaximize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -123,28 +126,28 @@ class PuLPTest(unittest.TestCase):
         print("\t Testing unbounded continuous LP solution")
         if self.solver.__class__ in [GUROBI, CPLEX_CMD, YAPOSIB, CPLEX_PY, MOSEK]:
             # These solvers report infeasible or unbounded
-            pulpTestCheck(prob, self.solver, [LpStatusInfeasible, LpStatusUnbounded])
+            pulpTestCheck(prob, self.solver, [const.LpStatusInfeasible, const.LpStatusUnbounded])
         elif self.solver.__class__ in [COINMP_DLL, MIPCL_CMD]:
             # COINMP_DLL is just plain wrong
             # also MIPCL_CMD
             print('\t\t Error in CoinMP and MIPCL_CMD: reports Optimal')
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal])
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal])
         elif self.solver.__class__ is GLPK_CMD:
             # GLPK_CMD Does not report unbounded problems, correctly
-            pulpTestCheck(prob, self.solver, [LpStatusUndefined])
+            pulpTestCheck(prob, self.solver, [const.LpStatusUndefined])
         elif self.solver.__class__ in [CPLEX_DLL, GUROBI_CMD]:
             # CPLEX_DLL Does not report unbounded problems, correctly
             # GUROBI_CMD has a very simple interface
-            pulpTestCheck(prob, self.solver, [LpStatusNotSolved])
+            pulpTestCheck(prob, self.solver, [const.LpStatusNotSolved])
         elif self.solver.__class__ in [PULP_CHOCO_CMD, CHOCO_CMD]:
             # choco bounds all variables. Would not return unbounded status
             pass
         else:
-            pulpTestCheck(prob, self.solver, [LpStatusUnbounded])
+            pulpTestCheck(prob, self.solver, [const.LpStatusUnbounded])
 
     def test_pulp_013(self):
         # Long name
-        prob = LpProblem("test013", LpMinimize)
+        prob = LpProblem("test013", const.LpMinimize)
         x = LpVariable("x" * 120, 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -157,16 +160,16 @@ class PuLPTest(unittest.TestCase):
         print("\t Testing Long Names")
         if self.solver.__class__ in [CPLEX_CMD, GLPK_CMD, GUROBI_CMD, MIPCL_CMD]:
             try:
-                pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
+                pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
             except PulpError:
                 # these solvers should raise an error'
                 pass
         else:
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
 
     def test_pulp_014(self):
         # repeated name
-        prob = LpProblem("test014", LpMinimize)
+        prob = LpProblem("test014", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("x", -1, 1)
         z = LpVariable("z", 0)
@@ -181,16 +184,16 @@ class PuLPTest(unittest.TestCase):
                                 CPLEX_CMD, CPLEX_DLL, CPLEX_PY,
                                 GLPK_CMD, GUROBI_CMD, PULP_CHOCO_CMD, CHOCO_CMD, MIPCL_CMD, MOSEK]:
             try:
-                pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
+                pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
             except PulpError:
                 # these solvers should raise an error
                 pass
         else:
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
 
     def test_pulp_015(self):
         # zero constraint
-        prob = LpProblem("test015", LpMinimize)
+        prob = LpProblem("test015", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -202,11 +205,11 @@ class PuLPTest(unittest.TestCase):
         prob += w >= 0, "c4"
         prob += lpSum([0, 0]) <= 0, "c5"
         print("\t Testing zero constraint")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
 
     def test_pulp_016(self):
         # zero objective
-        prob = LpProblem("test016", LpMinimize)
+        prob = LpProblem("test016", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -217,11 +220,11 @@ class PuLPTest(unittest.TestCase):
         prob += w >= 0, "c4"
         prob += lpSum([0, 0]) <= 0, "c5"
         print("\t Testing zero objective")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal])
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal])
 
     def test_pulp_017(self):
         # variable as objective
-        prob = LpProblem("test017", LpMinimize)
+        prob = LpProblem("test017", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -233,11 +236,11 @@ class PuLPTest(unittest.TestCase):
         prob += w >= 0, "c4"
         prob += lpSum([0, 0]) <= 0, "c5"
         print("\t Testing LpVariable (not LpAffineExpression) objective")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal])
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal])
 
     def test_pulp_018(self):
         # Long name in lp
-        prob = LpProblem("test018", LpMinimize)
+        prob = LpProblem("test018", const.LpMinimize)
         x = LpVariable("x" * 90, 0, 4)
         y = LpVariable("y" * 90, -1, 1)
         z = LpVariable("z" * 90, 0)
@@ -249,12 +252,12 @@ class PuLPTest(unittest.TestCase):
         prob += w >= 0, "c4"
         if self.solver.__class__ in [PULP_CBC_CMD, COIN_CMD]:
             print("\t Testing Long lines in LP")
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0},
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0},
                           use_mps=False)
 
     def test_pulp_019(self):
         # divide
-        prob = LpProblem("test019", LpMinimize)
+        prob = LpProblem("test019", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -265,41 +268,41 @@ class PuLPTest(unittest.TestCase):
         prob += -y + z == 7, "c3"
         prob += w >= 0, "c4"
         print("\t Testing LpAffineExpression divide")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6, w: 0})
 
     def test_pulp_020(self):
         # MIP
-        prob = LpProblem("test020", LpMinimize)
+        prob = LpProblem("test020", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
-        z = LpVariable("z", 0, None, LpInteger)
+        z = LpVariable("z", 0, None, const.LpInteger)
         prob += x + 4 * y + 9 * z, "obj"
         prob += x + y <= 5, "c1"
         prob += x + z >= 10, "c2"
         prob += -y + z == 7.5, "c3"
         print("\t Testing MIP solution")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 3, y: -0.5, z: 7})
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 3, y: -0.5, z: 7})
 
     def test_pulp_021(self):
         # MIP with floats in objective
-        prob = LpProblem("test021", LpMinimize)
+        prob = LpProblem("test021", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
-        z = LpVariable("z", 0, None, LpInteger)
+        z = LpVariable("z", 0, None, const.LpInteger)
         prob += 1.1 * x + 4.1 * y + 9.1 * z, "obj"
         prob += x + y <= 5, "c1"
         prob += x + z >= 10, "c2"
         prob += -y + z == 7.5, "c3"
         print("\t Testing MIP solution with floats in objective")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 3, y: -0.5, z: 7},
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 3, y: -0.5, z: 7},
                       objective=64.95)
 
     def test_pulp_030(self):
         # relaxed MIP
-        prob = LpProblem("test030", LpMinimize)
+        prob = LpProblem("test030", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
-        z = LpVariable("z", 0, None, LpInteger)
+        z = LpVariable("z", 0, None, const.LpInteger)
         prob += x + 4 * y + 9 * z, "obj"
         prob += x + y <= 5, "c1"
         prob += x + z >= 10, "c2"
@@ -308,25 +311,25 @@ class PuLPTest(unittest.TestCase):
         print("\t Testing MIP relaxation")
         if self.solver.__class__ in [GUROBI_CMD, PULP_CHOCO_CMD, CHOCO_CMD, MIPCL_CMD]:
             # gurobi command, choco and mipcl do not let the problem be relaxed
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 3.0, y: -0.5, z: 7})
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 3.0, y: -0.5, z: 7})
         else:
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 3.5, y: -1, z: 6.5})
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 3.5, y: -1, z: 6.5})
 
     def test_pulp_040(self):
         # Feasibility only
-        prob = LpProblem("test040", LpMinimize)
+        prob = LpProblem("test040", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
-        z = LpVariable("z", 0, None, LpInteger)
+        z = LpVariable("z", 0, None, const.LpInteger)
         prob += x + y <= 5, "c1"
         prob += x + z >= 10, "c2"
         prob += -y + z == 7.5, "c3"
         print("\t Testing feasibility problem (no objective)")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal])
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal])
 
     def test_pulp_050(self):
         # Infeasible
-        prob = LpProblem("test050", LpMinimize)
+        prob = LpProblem("test050", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0, 10)
@@ -336,83 +339,83 @@ class PuLPTest(unittest.TestCase):
         print("\t Testing an infeasible problem")
         if self.solver.__class__ is GLPK_CMD:
             # GLPK_CMD return codes are not informative enough
-            pulpTestCheck(prob, self.solver, [LpStatusUndefined])
+            pulpTestCheck(prob, self.solver, [const.LpStatusUndefined])
         elif self.solver.__class__ in [CPLEX_DLL, GUROBI_CMD]:
             # CPLEX_DLL Does not solve the problem
-            pulpTestCheck(prob, self.solver, [LpStatusNotSolved])
+            pulpTestCheck(prob, self.solver, [const.LpStatusNotSolved])
         else:
-            pulpTestCheck(prob, self.solver, [LpStatusInfeasible])
+            pulpTestCheck(prob, self.solver, [const.LpStatusInfeasible])
 
     def test_pulp_060(self):
         # Integer Infeasible
-        prob = LpProblem("test060", LpMinimize)
-        x = LpVariable("x", 0, 4, LpInteger)
-        y = LpVariable("y", -1, 1, LpInteger)
-        z = LpVariable("z", 0, 10, LpInteger)
+        prob = LpProblem("test060", const.LpMinimize)
+        x = LpVariable("x", 0, 4, const.LpInteger)
+        y = LpVariable("y", -1, 1, const.LpInteger)
+        z = LpVariable("z", 0, 10, const.LpInteger)
         prob += x + y <= 5.2, "c1"
         prob += x + z >= 10.3, "c2"
         prob += -y + z == 7.4, "c3"
         print("\t Testing an integer infeasible problem")
         if self.solver.__class__ in [GLPK_CMD, COIN_CMD, PULP_CBC_CMD, MOSEK]:
             # GLPK_CMD returns InfeasibleOrUnbounded
-            pulpTestCheck(prob, self.solver, [LpStatusInfeasible, LpStatusUndefined])
+            pulpTestCheck(prob, self.solver, [const.LpStatusInfeasible, const.LpStatusUndefined])
         elif self.solver.__class__ in [COINMP_DLL]:
             # Currently there is an error in COINMP for problems where
             # presolve eliminates too many variables
             print("\t\t Error in CoinMP to be fixed, reports Optimal")
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal])
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal])
         elif self.solver.__class__ in [GUROBI_CMD]:
-            pulpTestCheck(prob, self.solver, [LpStatusNotSolved])
+            pulpTestCheck(prob, self.solver, [const.LpStatusNotSolved])
         else:
-            pulpTestCheck(prob, self.solver, [LpStatusInfeasible])
+            pulpTestCheck(prob, self.solver, [const.LpStatusInfeasible])
 
     def test_pulp_070(self):
         # Column Based modelling of test_pulp_1
-        prob = LpProblem("test070", LpMinimize)
+        prob = LpProblem("test070", const.LpMinimize)
         obj = LpConstraintVar("obj")
         # constraints
-        a = LpConstraintVar("C1", LpConstraintLE, 5)
-        b = LpConstraintVar("C2", LpConstraintGE, 10)
-        c = LpConstraintVar("C3", LpConstraintEQ, 7)
+        a = LpConstraintVar("C1", const.LpConstraintLE, 5)
+        b = LpConstraintVar("C2", const.LpConstraintGE, 10)
+        c = LpConstraintVar("C3", const.LpConstraintEQ, 7)
 
         prob.setObjective(obj)
         prob += a
         prob += b
         prob += c
         # Variables
-        x = LpVariable("x", 0, 4, LpContinuous, obj + a + b)
-        y = LpVariable("y", -1, 1, LpContinuous, 4 * obj + a - c)
-        z = LpVariable("z", 0, None, LpContinuous, 9 * obj + b + c)
+        x = LpVariable("x", 0, 4, const.LpContinuous, obj + a + b)
+        y = LpVariable("y", -1, 1, const.LpContinuous, 4 * obj + a - c)
+        z = LpVariable("z", 0, None, const.LpContinuous, 9 * obj + b + c)
         print("\t Testing column based modelling")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6})
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6})
 
     def test_pulp_075(self):
         # Column Based modelling of test_pulp_1 with empty constraints
-        prob = LpProblem("test075", LpMinimize)
+        prob = LpProblem("test075", const.LpMinimize)
         obj = LpConstraintVar("obj")
         # constraints
-        a = LpConstraintVar("C1", LpConstraintLE, 5)
-        b = LpConstraintVar("C2", LpConstraintGE, 10)
-        c = LpConstraintVar("C3", LpConstraintEQ, 7)
+        a = LpConstraintVar("C1", const.LpConstraintLE, 5)
+        b = LpConstraintVar("C2", const.LpConstraintGE, 10)
+        c = LpConstraintVar("C3", const.LpConstraintEQ, 7)
 
         prob.setObjective(obj)
         prob += a
         prob += b
         prob += c
         # Variables
-        x = LpVariable("x", 0, 4, LpContinuous, obj + b)
-        y = LpVariable("y", -1, 1, LpContinuous, 4 * obj - c)
-        z = LpVariable("z", 0, None, LpContinuous, 9 * obj + b + c)
+        x = LpVariable("x", 0, 4, const.LpContinuous, obj + b)
+        y = LpVariable("y", -1, 1, const.LpContinuous, 4 * obj - c)
+        z = LpVariable("z", 0, None, const.LpContinuous, 9 * obj + b + c)
         if self.solver.__class__ in [CPLEX_DLL, CPLEX_CMD, COINMP_DLL, YAPOSIB,
                                 PYGLPK]:
             print("\t Testing column based modelling with empty constraints")
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal], {x: 4, y: -1, z: 6})
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal], {x: 4, y: -1, z: 6})
 
     def test_pulp_080(self):
         """
         Test the reporting of dual variables slacks and reduced costs
         """
-        prob = LpProblem("test080", LpMinimize)
+        prob = LpProblem("test080", const.LpMinimize)
         x = LpVariable("x", 0, 5)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -428,7 +431,7 @@ class PuLPTest(unittest.TestCase):
         if self.solver.__class__ in [CPLEX_DLL, CPLEX_CMD, COINMP_DLL,
                                 PULP_CBC_CMD, YAPOSIB, PYGLPK]:
             print("\t Testing dual variables and slacks reporting")
-            pulpTestCheck(prob, self.solver, [LpStatusOptimal],
+            pulpTestCheck(prob, self.solver, [const.LpStatusOptimal],
                           sol={x: 4, y: -1, z: 6},
                           reducedcosts={x: 0, y: 12, z: 0},
                           duals={"c1": 0, "c2": 1, "c3": 8},
@@ -436,12 +439,12 @@ class PuLPTest(unittest.TestCase):
 
     def test_pulp_090(self):
         # Column Based modelling of test_pulp_1 with a resolve
-        prob = LpProblem("test090", LpMinimize)
+        prob = LpProblem("test090", const.LpMinimize)
         obj = LpConstraintVar("obj")
         # constraints
-        a = LpConstraintVar("C1", LpConstraintLE, 5)
-        b = LpConstraintVar("C2", LpConstraintGE, 10)
-        c = LpConstraintVar("C3", LpConstraintEQ, 7)
+        a = LpConstraintVar("C1", const.LpConstraintLE, 5)
+        b = LpConstraintVar("C2", const.LpConstraintGE, 10)
+        c = LpConstraintVar("C3", const.LpConstraintEQ, 7)
 
         prob.setObjective(obj)
         prob += a
@@ -449,23 +452,23 @@ class PuLPTest(unittest.TestCase):
         prob += c
 
         prob.setSolver(self.solver)  # Variables
-        x = LpVariable("x", 0, 4, LpContinuous, obj + a + b)
-        y = LpVariable("y", -1, 1, LpContinuous, 4 * obj + a - c)
+        x = LpVariable("x", 0, 4, const.LpContinuous, obj + a + b)
+        y = LpVariable("y", -1, 1, const.LpContinuous, 4 * obj + a - c)
         prob.resolve()
-        z = LpVariable("z", 0, None, LpContinuous, 9 * obj + b + c)
+        z = LpVariable("z", 0, None, const.LpContinuous, 9 * obj + b + c)
         if self.solver.__class__ in [CPLEX_DLL, COINMP_DLL]:
             print("\t Testing resolve of problem")
             prob.resolve()
             # difficult to check this is doing what we want as the resolve is
             # over ridden if it is not implemented
-            # test_pulp_Check(prob, self.solver, [LpStatusOptimal], {x:4, y:-1, z:6})
+            # test_pulp_Check(prob, self.solver, [const.LpStatusOptimal], {x:4, y:-1, z:6})
 
     def test_pulp_100(self):
         """
         Test the ability to sequentially solve a problem
         """
         # set up a cubic feasible region
-        prob = LpProblem("test100", LpMinimize)
+        prob = LpProblem("test100", const.LpMinimize)
         x = LpVariable("x", 0, 1)
         y = LpVariable("y", 0, 1)
         z = LpVariable("z", 0, 1)
@@ -477,7 +480,7 @@ class PuLPTest(unittest.TestCase):
         if self.solver.__class__ in [CPLEX_DLL, COINMP_DLL, GUROBI]:
             print("\t Testing Sequential Solves")
             status = prob.sequentialSolve([obj1, obj2], solver=self.solver)
-            pulpTestCheck(prob, self.solver, [[LpStatusOptimal, LpStatusOptimal]],
+            pulpTestCheck(prob, self.solver, [[const.LpStatusOptimal, const.LpStatusOptimal]],
                           sol={x: 0, y: 1},
                           status=status)
 
@@ -485,7 +488,7 @@ class PuLPTest(unittest.TestCase):
         """
         Test the ability to use fractional constraints
         """
-        prob = LpProblem("test110", LpMinimize)
+        prob = LpProblem("test110", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -495,16 +498,16 @@ class PuLPTest(unittest.TestCase):
         prob += x + z >= 10, "c2"
         prob += -y + z == 7, "c3"
         prob += w >= 0, "c4"
-        prob += LpFractionConstraint(x, z, LpConstraintEQ, 0.5, name='c5')
+        prob += LpFractionConstraint(x, z, const.LpConstraintEQ, 0.5, name='c5')
         print("\t Testing fractional constraints")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal],
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal],
                       {x: 10 / 3.0, y: -1 / 3.0, z: 20 / 3.0, w: 0})
 
     def test_pulp_120(self):
         """
         Test the ability to use Elastic constraints
         """
-        prob = LpProblem("test120", LpMinimize)
+        prob = LpProblem("test120", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -515,14 +518,14 @@ class PuLPTest(unittest.TestCase):
         prob += -y + z == 7, "c3"
         prob.extend((w >= -1).makeElasticSubProblem())
         print("\t Testing elastic constraints (no change)")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal],
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal],
                       {x: 4, y: -1, z: 6, w: -1})
 
     def test_pulp_121(self):
         """
         Test the ability to use Elastic constraints
         """
-        prob = LpProblem("test121", LpMinimize)
+        prob = LpProblem("test121", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -533,14 +536,14 @@ class PuLPTest(unittest.TestCase):
         prob += -y + z == 7, "c3"
         prob.extend((w >= -1).makeElasticSubProblem(proportionFreeBound=0.1))
         print("\t Testing elastic constraints (freebound)")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal],
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal],
                       {x: 4, y: -1, z: 6, w: -1.1})
 
     def test_pulp_122(self):
         """
         Test the ability to use Elastic constraints (penalty unchanged)
         """
-        prob = LpProblem("test122", LpMinimize)
+        prob = LpProblem("test122", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -551,14 +554,14 @@ class PuLPTest(unittest.TestCase):
         prob += -y + z == 7, "c3"
         prob.extend((w >= -1).makeElasticSubProblem(penalty=1.1))
         print("\t Testing elastic constraints (penalty unchanged)")
-        pulpTestCheck(prob, self.solver, [LpStatusOptimal],
+        pulpTestCheck(prob, self.solver, [const.LpStatusOptimal],
                       {x: 4, y: -1, z: 6, w: -1.0})
 
     def test_pulp_123(self):
         """
         Test the ability to use Elastic constraints (penalty unbounded)
         """
-        prob = LpProblem("test123", LpMinimize)
+        prob = LpProblem("test123", const.LpMinimize)
         x = LpVariable("x", 0, 4)
         y = LpVariable("y", -1, 1)
         z = LpVariable("z", 0)
@@ -571,18 +574,18 @@ class PuLPTest(unittest.TestCase):
         print("\t Testing elastic constraints (penalty unbounded)")
         if self.solver.__class__ in [COINMP_DLL, GUROBI, CPLEX_CMD, CPLEX_PY, YAPOSIB, MOSEK]:
             # COINMP_DLL Does not report unbounded problems, correctly
-            pulpTestCheck(prob, self.solver, [LpStatusInfeasible, LpStatusUnbounded])
+            pulpTestCheck(prob, self.solver, [const.LpStatusInfeasible, const.LpStatusUnbounded])
         elif self.solver.__class__ is GLPK_CMD:
             # GLPK_CMD Does not report unbounded problems, correctly
-            pulpTestCheck(prob, self.solver, [LpStatusUndefined])
+            pulpTestCheck(prob, self.solver, [const.LpStatusUndefined])
         elif self.solver.__class__ in [CPLEX_DLL, GUROBI_CMD]:
             # GLPK_CMD Does not report unbounded problems, correctly
-            pulpTestCheck(prob, self.solver, [LpStatusNotSolved])
+            pulpTestCheck(prob, self.solver, [const.LpStatusNotSolved])
         elif self.solver.__class__ in [PULP_CHOCO_CMD, CHOCO_CMD]:
             # choco bounds all variables. Would not return unbounded status
             pass
         else:
-            pulpTestCheck(prob, self.solver, [LpStatusUnbounded])
+            pulpTestCheck(prob, self.solver, [const.LpStatusUnbounded])
 
 
 def pulpTestCheck(prob, solver, okstatus, sol=None,
@@ -598,8 +601,8 @@ def pulpTestCheck(prob, solver, okstatus, sol=None,
     if status not in okstatus:
         dumpTestProblem(prob)
         print("Failure: status ==", status, "not in", okstatus)
-        print("Failure: status ==", LpStatus[status], "not in", \
-              [LpStatus[s] for s in okstatus])
+        print("Failure: status ==", const.LpStatus[status], "not in", \
+              [const.LpStatus[s] for s in okstatus])
         raise PulpError("Tests failed for solver %s" % solver)
     if sol is not None:
         for v, x in sol.items():
@@ -658,8 +661,11 @@ def suite():
     suite = unittest.TestSuite()
     for solver in solvers:
         if solver().available():
+            print("Solver %s available" % solver)
             tests = loader.loadTestsFromTestCase(PuLPTest, solver=solver(msg=0))
             suite.addTests(tests)
+        else:
+            print("Solver %s unavailable" % solver)
     return suite
